Fix split package in analysis-nori plugin
The analysis-nori plugin reuses server the server package name for
analysis. This commit moves the plugin implementation to use a single
package name, o.e.p.analysis.nori
