+++ plugins/analysis-nori/src/test/resources/org/elasticsearch/plugin/analysis/nori/user_dict.txt
